third_party/blink/web_tests/fast/selectors/adjacent-selectors-with-subselector-expected.txt

Test if direct adjacent selectors are counted correctly where combined with subselector.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(target, null).color is "rgb(0, 0, 0)"
PASS getComputedStyle(target, null).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE

This text should be green.
